Move inline spec modules into its own namespace
