+++ spec/integration/union_spec.rb
@@ -1,81 +1,14 @@
 # frozen_string_literal: true
 
 RSpec.describe Dry::Struct::Union do
-  module Solar
-    class Base < Dry::Struct
-      abstract
-      schema schema.strict
-    end
-
-    module Types
-      include Dry::Types()
-    end
-
-    module Weather
-      include Dry::Struct.Union(include: [:Warm])
-      MAX_TEMP = 274
-
-      class Base < Solar::Base
-        abstract
-        attribute :temp, "integer"
-      end
-
-      class Cold < Base
-        attribute :id, Types.Value(:cold)
-      end
-
-      class Warm < Base
-        attribute :id, Types.Value(:warm)
-      end
-    end
-
-    module Season
-      include Dry::Struct::Union
-
-      class Spring < Solar::Base
-        attribute :id, Types.Value(:spring)
-      end
-
-      module Unused
-        class Autum < Solar::Base
-          attribute :id, Types.Value(:autum)
-        end
-      end
-    end
-
-    module Planet
-      include Dry::Struct.Union(exclude: :Pluto)
-
-      class Base < Solar::Base
-        abstract
-        attribute? :closest, Planet
-        attribute? :season, Season
-        attribute? :weather, Weather
-      end
-
-      class Pluto < Base
-        attribute :id, Types.Value(:pluto)
-      end
-
-      class Earth < Base
-        attribute :id, Types.Value(:earth)
-      end
-
-      class Mars < Base
-        attribute :id, Types.Value(:mars)
-      end
-    end
-  end
-
-  describe Solar do
-    let(:remove) { "" }
+  describe Union::Example do
     let(:cold) { {id: :cold, temp: 100} }
     let(:warm) { {id: :warm, temp: 100} }
     let(:spring) { {id: :spring} }
 
     describe described_class::Weather do
       describe ".name" do
-        it { is_expected.to have_attributes(name: "#{remove}Solar::Weather<[#{remove}Solar::Weather::Warm]>") }
+        it { is_expected.to have_attributes(name: "Union::Example::Weather<[Union::Example::Weather::Warm]>") }
       end
 
       describe ".call" do
@@ -99,7 +32,7 @@ RSpec.describe Dry::Struct::Union do
       let(:mars) { {id: :mars, closest: earth, weather: warm} }
 
       describe ".name" do
-        it { is_expected.to have_attributes(name: "#{remove}Solar::Planet<[#{remove}Solar::Planet::Earth | #{remove}Solar::Planet::Mars]>") }
+        it { is_expected.to have_attributes(name: "Union::Example::Planet<[Union::Example::Planet::Earth | Union::Example::Planet::Mars]>") }
       end
 
       describe ".call" do
@@ -127,7 +60,7 @@ RSpec.describe Dry::Struct::Union do
 
     describe described_class::Season do
       describe ".name" do
-        it { is_expected.to have_attributes(name: "#{remove}Solar::Season<[#{remove}Solar::Season::Spring]>") }
+        it { is_expected.to have_attributes(name: "Union::Example::Season<[Union::Example::Season::Spring]>") }
       end
 
       describe ".call" do
