+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@
 
 require_relative "support/coverage"
 require_relative "support/warnings"
+require_relative "support/union"
 
 require "pathname"
